fix(reporter-junit): move junit test data to own file
